Refactoring: test revamp, moved happiness messages to the OrderPaidSetter to remove OrderShipper in future
